+++ packages/astro/src/runtime/ssr.ts
@@ -9,8 +9,8 @@ import { fileURLToPath } from 'url';
 import fs from 'fs';
 import path from 'path';
 import slash from 'slash';
-import { renderPage } from './astro.js';
 import glob from 'tiny-glob';
+import { renderPage } from '../internal/index.js';
 import { generatePaginateFunction } from './paginate.js';
 import { getParams, validateGetStaticPathsModule, validateGetStaticPathsResult } from './routing.js';
 import { parseNpmName, canonicalURL as getCanonicalURL, codeFrame } from './util.js';
